requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dnspython==2.1.0
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
LinkChecker==10.0.1
packaging==20.9
pbr==5.5.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
